pyproject.toml
[tool.poetry]
name = "execution-service"
description = ""
readme = "README.md"
package-mode = false

[tool.pytest.ini_options]
addopts = "-v --tb=short"
env_files = [".env/.pytest"]


[tool.poetry.dependencies]
python = "^3.11"
aws-requests-auth = "0.4.3"
aws-lambda-powertools = ">=2.15.0,<3"
boto3 = ">=1.28.80,<2.0.0"
botocore = "1.31.80"
boto3-stubs = {version = ">=1.26.115,<2", extras = ["batch", "dynamodb", "ec2", "events", "iam", "lambda", "logs", "sqs"]}
jinja2 = "3.0.3"
pydantic = ">=1.8.2,<2"
python-dotenv = "0.17.1"
pyyaml = "6.0.1"
slack-sdk = "3.17.0"
jsonpath-ng = "1.5.3"
cachetools = "5.2.0"
typing-extensions = "4.5.0"
google-cloud-logging = "3.8.0"
google-cloud-batch = "0.17.4"
google-cloud-kms = "2.19.2"
google-auth = "2.23.4"
google-auth-httplib2 = "0.1.1"
jit-utils = { git = "ssh://git@github.com/jitsecurity/jit-utils.git", rev = "36.1.0", extras = ["all"] }
py-api = { git = "ssh://git@github.com/jitsecurity/py_api.git", rev = "v0.1.6" }


[tool.poetry.group.dev.dependencies]
flake8 = "3.9.2"
fastapi = "0.65.2"
moto = "4.1.3"
freezegun = "1.2.2"
pip-licenses = "3.5.3"
pytest = ">=6.2.4,<8"
pytest-cov = "4.0.0"
pytest-dotenv = "0.5.2"
pytest-mock = "3.6.0"
requests-mock = "1.9.2"
pytest-xdist = ">=3.3.1"
uvicorn = "0.18.1"
responses = "0.23.1"
pydantic-factories = "1.1.0"
pre-commit = ">=3.3,<4"
test-utils = { git = "ssh://git@github.com/jitsecurity/test-utils.git", rev = "1.1.4" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
